requirements.txt
Flask 
Flask-Cors==3.0.10
gunicorn==20.1.0
pandas
mlxtend
numpy
python-dotenv
datetime
sqlalchemy
pymysql
pandasql